+++ src/PSRule.Rules.Azure/rules/Azure.Storage.Rule.ps1
@@ -9,7 +9,7 @@ Rule 'Azure.Storage.UseReplication' -If { ResourceType 'Microsoft.Storage/storag
 
 # Synopsis: Storage accounts should only accept secure traffic
 Rule 'Azure.Storage.SecureTransferRequired' -If { ResourceType 'Microsoft.Storage/storageAccounts' } -Tag @{ severity = 'Important'; category = 'Security configuration' } {
-    $TargetObject.Properties.supportsHttpsTrafficOnly
+    $TargetObject.Properties.supportsHttpsTrafficOnly -eq $True
 }
 
 # Synopsis: Storage Service Encryption (SSE) should be enabled
@@ -17,7 +17,10 @@ Rule 'Azure.Storage.UseEncryption' -If { ResourceType 'Microsoft.Storage/storage
     ($Null -ne $TargetObject.Properties.encryption) -and
     ($Null -ne $TargetObject.Properties.encryption.services.blob) -and
     ($Null -ne $TargetObject.Properties.encryption.services.file) -and
-    ($TargetObject.Properties.encryption.services.blob.enabled -and $TargetObject.Properties.encryption.services.file.enabled)
+    (
+        ($TargetObject.Properties.encryption.services.blob.enabled -eq $True) -and
+        ($TargetObject.Properties.encryption.services.file.enabled -eq $True)
+    )
 }
 
 # Synopsis: Enable soft delete on Storage Accounts
@@ -25,5 +28,5 @@ Rule 'Azure.Storage.SoftDelete' -If { ResourceType 'Microsoft.Storage/storageAcc
     $serviceProperties = $TargetObject.resources | Where-Object -FilterScript {
         $_.ResourceType -eq 'serviceProperties'
     }
-    $serviceProperties.DeleteRetentionPolicy.Enabled
+    $serviceProperties.DeleteRetentionPolicy.Enabled -eq $True
 }