+++ tests/PSRule.Rules.Azure.Tests/Azure.MySQL.Tests.ps1
@@ -24,7 +24,7 @@ Describe 'Azure.MySQL' {
     $dataPath = Join-Path -Path $here -ChildPath 'Resources.MySQL.json';
 
     Context 'Conditions' {
-        $result = Invoke-PSRule -Module PSRule.Rules.Azure -InputPath $dataPath -WarningAction Ignore;
+        $result = Invoke-PSRule -Module PSRule.Rules.Azure -InputPath $dataPath -WarningAction Ignore -ErrorAction Stop;
 
         It 'Azure.MySQL.UseSSL' {
             $filteredResult = $result | Where-Object { $_.RuleName -eq 'Azure.MySQL.UseSSL' };